+++ modules/AWSPowerShell/Cmdlets/CloudFront/AWS.Tools.CloudFront.Completers.psm1
@@ -531,6 +531,7 @@ $CF_SelectMap = @{
                "Update-CFCloudFrontOriginAccessIdentity",
                "Update-CFContinuousDeploymentPolicy",
                "Update-CFDistribution",
+               "Update-CFDistributionWithStagingConfig",
                "Update-CFFieldLevelEncryptionConfig",
                "Update-CFFieldLevelEncryptionProfile",
                "Update-CFFunction",