+++ modules/AWSPowerShell/Cmdlets/EKS/AWS.Tools.EKS.Completers.psm1
@@ -196,6 +196,7 @@ $EKS_SelectMap = @{
                "Remove-EKSNodegroup",
                "Unregister-EKSCluster",
                "Get-EKSAddon",
+               "Get-EKSAddonConfiguration",
                "Get-EKSAddonVersion",
                "Get-EKSCluster",
                "Get-EKSFargateProfile",