+++ modules/AWSPowerShell/Cmdlets/EKS/Basic/Get-EKSAddonConfiguration-Cmdlet.cs
@@ -1,0 +1,208 @@
+/*******************************************************************************
+ *  Copyright 2012-2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use
+ *  this file except in compliance with the License. A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ *  or in the "license" file accompanying this file.
+ *  This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ *  CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations under the License.
+ * *****************************************************************************
+ *
+ *  AWS Tools for Windows (TM) PowerShell (TM)
+ *
+ */
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Management.Automation;
+using System.Text;
+using Amazon.PowerShell.Common;
+using Amazon.Runtime;
+using Amazon.EKS;
+using Amazon.EKS.Model;
+
+namespace Amazon.PowerShell.Cmdlets.EKS
+{
+    /// <summary>
+    /// Returns configuration options.
+    /// </summary>
+    [Cmdlet("Get", "EKSAddonConfiguration")]
+    [OutputType("Amazon.EKS.Model.DescribeAddonConfigurationResponse")]
+    [AWSCmdlet("Calls the Amazon Elastic Container Service for Kubernetes DescribeAddonConfiguration API operation.", Operation = new[] {"DescribeAddonConfiguration"}, SelectReturnType = typeof(Amazon.EKS.Model.DescribeAddonConfigurationResponse))]
+    [AWSCmdletOutput("Amazon.EKS.Model.DescribeAddonConfigurationResponse",
+        "This cmdlet returns an Amazon.EKS.Model.DescribeAddonConfigurationResponse object containing multiple properties. The object can also be referenced from properties attached to the cmdlet entry in the $AWSHistory stack."
+    )]
+    public partial class GetEKSAddonConfigurationCmdlet : AmazonEKSClientCmdlet, IExecutor
+    {
+        
+        #region Parameter AddonName
+        /// <summary>
+        /// <para>
+        /// <para>The name of the add-on. The name must match one of the names returned by <a href="https://docs.aws.amazon.com/eks/latest/APIReference/API_DescribeAddonVersions.html"><code>DescribeAddonVersions</code></a>.</para>
+        /// </para>
+        /// </summary>
+        #if !MODULAR
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        #else
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true, Mandatory = true)]
+        [System.Management.Automation.AllowEmptyString]
+        [System.Management.Automation.AllowNull]
+        #endif
+        [Amazon.PowerShell.Common.AWSRequiredParameter]
+        public System.String AddonName { get; set; }
+        #endregion
+        
+        #region Parameter AddonVersion
+        /// <summary>
+        /// <para>
+        /// <para>The version of the add-on. The version must match one of the versions returned by
+        /// <a href="https://docs.aws.amazon.com/eks/latest/APIReference/API_DescribeAddonVersions.html"><code>DescribeAddonVersions</code></a>.</para>
+        /// </para>
+        /// </summary>
+        #if !MODULAR
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        #else
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true, Mandatory = true)]
+        [System.Management.Automation.AllowEmptyString]
+        [System.Management.Automation.AllowNull]
+        #endif
+        [Amazon.PowerShell.Common.AWSRequiredParameter]
+        public System.String AddonVersion { get; set; }
+        #endregion
+        
+        #region Parameter Select
+        /// <summary>
+        /// Use the -Select parameter to control the cmdlet output. The default value is '*'.
+        /// Specifying -Select '*' will result in the cmdlet returning the whole service response (Amazon.EKS.Model.DescribeAddonConfigurationResponse).
+        /// Specifying the name of a property of type Amazon.EKS.Model.DescribeAddonConfigurationResponse will result in that property being returned.
+        /// Specifying -Select '^ParameterName' will result in the cmdlet returning the selected cmdlet parameter value.
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public string Select { get; set; } = "*";
+        #endregion
+        
+        protected override void ProcessRecord()
+        {
+            base.ProcessRecord();
+            
+            var context = new CmdletContext();
+            
+            // allow for manipulation of parameters prior to loading into context
+            PreExecutionContextLoad(context);
+            
+            if (ParameterWasBound(nameof(this.Select)))
+            {
+                context.Select = CreateSelectDelegate<Amazon.EKS.Model.DescribeAddonConfigurationResponse, GetEKSAddonConfigurationCmdlet>(Select) ??
+                    throw new System.ArgumentException("Invalid value for -Select parameter.", nameof(this.Select));
+            }
+            context.AddonName = this.AddonName;
+            #if MODULAR
+            if (this.AddonName == null && ParameterWasBound(nameof(this.AddonName)))
+            {
+                WriteWarning("You are passing $null as a value for parameter AddonName which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
+            }
+            #endif
+            context.AddonVersion = this.AddonVersion;
+            #if MODULAR
+            if (this.AddonVersion == null && ParameterWasBound(nameof(this.AddonVersion)))
+            {
+                WriteWarning("You are passing $null as a value for parameter AddonVersion which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
+            }
+            #endif
+            
+            // allow further manipulation of loaded context prior to processing
+            PostExecutionContextLoad(context);
+            
+            var output = Execute(context) as CmdletOutput;
+            ProcessOutput(output);
+        }
+        
+        #region IExecutor Members
+        
+        public object Execute(ExecutorContext context)
+        {
+            var cmdletContext = context as CmdletContext;
+            // create request
+            var request = new Amazon.EKS.Model.DescribeAddonConfigurationRequest();
+            
+            if (cmdletContext.AddonName != null)
+            {
+                request.AddonName = cmdletContext.AddonName;
+            }
+            if (cmdletContext.AddonVersion != null)
+            {
+                request.AddonVersion = cmdletContext.AddonVersion;
+            }
+            
+            CmdletOutput output;
+            
+            // issue call
+            var client = Client ?? CreateClient(_CurrentCredentials, _RegionEndpoint);
+            try
+            {
+                var response = CallAWSServiceOperation(client, request);
+                object pipelineOutput = null;
+                pipelineOutput = cmdletContext.Select(response, this);
+                output = new CmdletOutput
+                {
+                    PipelineOutput = pipelineOutput,
+                    ServiceResponse = response
+                };
+            }
+            catch (Exception e)
+            {
+                output = new CmdletOutput { ErrorResponse = e };
+            }
+            
+            return output;
+        }
+        
+        public ExecutorContext CreateContext()
+        {
+            return new CmdletContext();
+        }
+        
+        #endregion
+        
+        #region AWS Service Operation Call
+        
+        private Amazon.EKS.Model.DescribeAddonConfigurationResponse CallAWSServiceOperation(IAmazonEKS client, Amazon.EKS.Model.DescribeAddonConfigurationRequest request)
+        {
+            Utils.Common.WriteVerboseEndpointMessage(this, client.Config, "Amazon Elastic Container Service for Kubernetes", "DescribeAddonConfiguration");
+            try
+            {
+                #if DESKTOP
+                return client.DescribeAddonConfiguration(request);
+                #elif CORECLR
+                return client.DescribeAddonConfigurationAsync(request).GetAwaiter().GetResult();
+                #else
+                        #error "Unknown build edition"
+                #endif
+            }
+            catch (AmazonServiceException exc)
+            {
+                var webException = exc.InnerException as System.Net.WebException;
+                if (webException != null)
+                {
+                    throw new Exception(Utils.Common.FormatNameResolutionFailureMessage(client.Config, webException.Message), webException);
+                }
+                throw;
+            }
+        }
+        
+        #endregion
+        
+        internal partial class CmdletContext : ExecutorContext
+        {
+            public System.String AddonName { get; set; }
+            public System.String AddonVersion { get; set; }
+            public System.Func<Amazon.EKS.Model.DescribeAddonConfigurationResponse, GetEKSAddonConfigurationCmdlet, object> Select { get; set; } =
+                (response, cmdlet) => response;
+        }
+        
+    }
+}