+++ modules/AWSPowerShell/Cmdlets/EKS/Basic/New-EKSAddon-Cmdlet.cs
@@ -103,6 +103,18 @@ namespace Amazon.PowerShell.Cmdlets.EKS
         public System.String ClusterName { get; set; }
         #endregion
         
+        #region Parameter ConfigurationValue
+        /// <summary>
+        /// <para>
+        /// <para> The set of configuration values for the add-on being created. Whatever values provided
+        /// here are validated against the schema from <a href="https://docs.aws.amazon.com/eks/latest/APIReference/API_DescribeAddonConfiguration.html"><code>DescribeAddonConfiguration</code></a>.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        [Alias("ConfigurationValues")]
+        public System.String ConfigurationValue { get; set; }
+        #endregion
+        
         #region Parameter ResolveConflict
         /// <summary>
         /// <para>
@@ -227,6 +239,7 @@ namespace Amazon.PowerShell.Cmdlets.EKS
                 WriteWarning("You are passing $null as a value for parameter ClusterName which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
             }
             #endif
+            context.ConfigurationValue = this.ConfigurationValue;
             context.ResolveConflict = this.ResolveConflict;
             context.ServiceAccountRoleArn = this.ServiceAccountRoleArn;
             if (this.Tag != null)
@@ -268,6 +281,10 @@ namespace Amazon.PowerShell.Cmdlets.EKS
             if (cmdletContext.ClusterName != null)
             {
                 request.ClusterName = cmdletContext.ClusterName;
+            }
+            if (cmdletContext.ConfigurationValue != null)
+            {
+                request.ConfigurationValues = cmdletContext.ConfigurationValue;
             }
             if (cmdletContext.ResolveConflict != null)
             {
@@ -346,6 +363,7 @@ namespace Amazon.PowerShell.Cmdlets.EKS
             public System.String AddonVersion { get; set; }
             public System.String ClientRequestToken { get; set; }
             public System.String ClusterName { get; set; }
+            public System.String ConfigurationValue { get; set; }
             public Amazon.EKS.ResolveConflicts ResolveConflict { get; set; }
             public System.String ServiceAccountRoleArn { get; set; }
             public Dictionary<System.String, System.String> Tag { get; set; }