+++ modules/AWSPowerShell/Cmdlets/EKS/Basic/New-EKSNodegroup-Cmdlet.cs
@@ -149,8 +149,8 @@ namespace Amazon.PowerShell.Cmdlets.EKS
         #region Parameter RemoteAccess_Ec2SshKey
         /// <summary>
         /// <para>
-        /// <para>The Amazon EC2 SSH key that provides access for SSH communication with the nodes in
-        /// the managed node group. For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-key-pairs.html">Amazon
+        /// <para>The Amazon EC2 SSH key name that provides access for SSH communication with the nodes
+        /// in the managed node group. For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-key-pairs.html">Amazon
         /// EC2 key pairs and Linux instances</a> in the <i>Amazon Elastic Compute Cloud User
         /// Guide for Linux Instances</i>.</para>
         /// </para>
@@ -325,9 +325,9 @@ namespace Amazon.PowerShell.Cmdlets.EKS
         #region Parameter RemoteAccess_SourceSecurityGroup
         /// <summary>
         /// <para>
-        /// <para>The security groups that are allowed SSH access (port 22) to the nodes. If you specify
-        /// an Amazon EC2 SSH key but do not specify a source security group when you create a
-        /// managed node group, then port 22 on the nodes is opened to the internet (0.0.0.0/0).
+        /// <para>The security group ids that are allowed SSH access (port 22) to the nodes. If you
+        /// specify an Amazon EC2 SSH key but do not specify a source security group when you
+        /// create a managed node group, then port 22 on the nodes is opened to the internet (0.0.0.0/0).
         /// For more information, see <a href="https://docs.aws.amazon.com/vpc/latest/userguide/VPC_SecurityGroups.html">Security
         /// Groups for Your VPC</a> in the <i>Amazon Virtual Private Cloud User Guide</i>.</para>
         /// </para>